g151 - Move file
Description
Move mobile related view to 'mobile' folder.
