+++ imports/stylesheets/partials/nav.import.less
@@ -49,8 +49,8 @@ nav {
         }
     }
     a{
-        height: 55px;
-        line-height: 55px;
+        height: 45px;
+        line-height: 45px;
     }
 
     .fa {
